+++ CHANGELOG.md
@@ -40,6 +40,7 @@
 - Module information and documentation could be viewed directly from the module list
 - Add the possibility to translate text in the sql files (insert.sql, update/sql/\*.sql). to generate sql files use command `php Thelia generate:sql`. Translation can be made in the back office, in the translation page.   
 - format_date smarty function now handle symfony form type ```date```, ```datetime``` and ```time``` view value.
+- Allow BaseController::generateOrderPdf to generate a pdf without having the rights
 
 # 2.1.3